update deserialization logic
