+++ packages/ensnode-sdk/src/indexing-status/mappers/serializer.ts
